requirements.txt
pytest
pytest-xdist
allure-pytest
faker
selenium
